Add files via upload
The program is actually not fully functional (its can run and will play out), but after I changed over all the enemies from just function calls to their own class and objects that's called in their own function, that also now has the loot tables linked to it, I couldn't figure out how to get it so the variable that needs to increase actually increases. But I'm just going to call it quits for now.